refactor magazine-footer to magazine-summary
