+++ blocks/magazine-summary/magazine-summary.css
@@ -1,0 +1,107 @@
+main .section.magazine-summary-container {
+  background-color: #f8f8f8;
+}
+
+.section div.magazine-summary-wrapper {
+  max-width: 1174px;
+  align-items: center;
+}
+
+.magazine-summary {
+  display: block;
+}
+
+.magazine-summary > div {
+  width: 100%;
+  padding: 0 10px;
+}
+
+.magazine-summary div.section-title {
+  text-transform: uppercase;
+  padding-top: 35px;
+  font-weight: 500;
+}
+
+.magazine-summary .col2 > *,
+.magazine-summary .col3 > * {
+  display: block;
+  padding: 6px 0;
+  line-height: 1.2em;
+  letter-spacing: 0.02em;
+}
+
+.magazine-summary h6.red {
+  color: #ee3124;
+}
+
+.magazine-summary picture {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 375px;
+}
+
+
+
+.magazine-summary img {
+  transform: rotate(-6deg);
+  box-shadow: -5px 5px 10px rgba(0 0 0 / 30%);
+  height: 300px;
+}
+
+@media (min-width: 900px) {
+  .magazine-summary {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+    max-width: 1174px;
+  }
+
+  .magazine-summary .col2,
+  .magazine-summary .col3 {
+    display: flex;
+    flex-direction: column;
+    align-self: flex-start;
+    padding-top: 100px;
+    width: 350px;
+  }
+
+  .magazine-summary.navigation .col1 {
+    text-align: right;
+    padding-bottom: 30px;
+  }
+
+  .magazine-summary.navigation .col1 h4 {
+    margin: 0;
+    padding-bottom: 20px;
+  }
+
+  .magazine-summary.navigation .col2,
+  .magazine-summary.navigation .col3 {
+    padding-top: 30px;
+    width: 300px;
+  }
+
+  .magazine-summary.navigation picture {
+    height: 275px;
+    padding-top: 50px;
+  }
+
+  .magazine-summary.navigation img {
+    height: 250px;
+  }
+
+  .magazine-summary.navigation a.button {
+    border: 2px solid #000;
+    padding: 10px 15px;
+    background-color: transparent;
+    color: var(--link-color);
+    width: 200px;
+    line-height: unset;
+  }
+
+  .magazine-summary.navigation a.button:hover {
+    background-color: var(--current-link-color);
+    color: var(--text-color-on-black-background);
+  }
+}
